+++ lib/spellr/cli.rb
@@ -13,26 +13,33 @@ module Spellr
     end
 
     def run
-      catch(:spellr_exit) { check }
+      catch(:spellr_exit) { run_subcommand }
     rescue Spellr::Error => e
-      Spellr.config.output.warn(Spellr::StringFormat.red(e.message)) && 1
+      Spellr.config.output.warn(Spellr::StringFormat.red(e.message))
       1
     end
 
     private
 
-    def check
+    def run_subcommand
       CLI::Options.parse(@argv)
       Spellr.config.valid?
-      checker = Spellr.config.checker.new(files: files)
+      exit_code = check
+      exit_code = prune if exit_code.zero? && Spellr.config.prune_wordlists?
+      exit_code
+    end
+
+    def check
+      checker = Spellr.config.checker.new(files: Spellr.config.file_list)
       checker.check
 
       checker.exit_code
     end
 
-    def files
-      require_relative 'file_list'
-      Spellr::FileList.new(@argv)
+    def prune
+      require_relative 'prune'
+      Spellr.config.output.puts ''
+      Spellr::Prune.run
     end
   end
 end